pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "unet-segmentation"
version = "0.1.0"
description = "U-Net модель для сегментации медицинских/архитектурных изображений"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Юлия", email = "enemydoctor@gmail.com"},
]
license = {text = "MIT"}

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "torchmetrics>=1.0.0",
    
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "albumentations>=1.3.0",
    "scikit-learn>=1.3.0",
    
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    
    "clearml>=1.10.0",
    
    "matplotlib>=3.7.0",
    
    "kagglehub>=0.1.0",
    
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["modules*", "configs*"]
namespaces = false

[project.scripts]
train-unet = "modules.train:main"

[tool.ruff.lint]
target-version = "py312"
select = [
    "E",
    "W",
    "F401",
    "I"
]
exclude = [
    ".git",
    "__pycache__",
    "data",
    "checkpoints",
    "configs",
    "unet-segmentation-egg.info",
    ".venv",
    "build",
    "dist"
]

[tool.ruff.format]

line-length = 88

[tool.isort]
profile = "black"
line_length = 88
skip = [
    "data",
    "checkpoints",
    ".venv",
    "build",
    "dist"
]